+++ Assets/Mizutani/Scripts/TextChange.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 04befa7590a693f4492f1ff732be9f38
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
