+++ Assets/Mizutani/Scripts/TextWriter.cs.meta
