+++ Assets/Nagasawa/Scenes/New Scene.unity
@@ -16906,7 +16906,7 @@ GameObject:
   m_Component:
   - component: {fileID: 2009754369}
   m_Layer: 0
-  m_Name: Clocks
+  m_Name: ClockNazoManager
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
